new file export process and buildsize testing

+++ Packages/com.pjkt.sdk/Editor/Systems/BoothUploader.cs
@@ -26,26 +26,43 @@ namespace PJKT.SDK2
                 PjktSdkWindow.Notify("Tried to upload a booth that doesn't exist. How did you do that?", BoothErrorType.Warning);
                 return;
             }
+            
+            PjktSdkWindow.Notify("Building...");
+
+            long buildsize = PjktBuildSize.AssessBuildSize(boothDescriptor);
+            long maxBuildSizeBytes = PjktEventManager.SelectedProjekt.booth_requirements.MaxBuildSize * 1024 * 1024;
+            if (buildsize > maxBuildSizeBytes)
+            {
+                PjktSdkWindow.Notify($"Booth exceeds maximum build size ({BoothValidator.FormatSize(buildsize)}/{PjktEventManager.SelectedProjekt.booth_requirements.MaxBuildSize})\nTry compressing your textures or meshes.\n Booth was not uploaded.", BoothErrorType.Error);
+                return;
+            }
 
             PjktSdkWindow.Notify("Uploading booth to server...");
-            
-            if (await CreateBoothPackage(boothDescriptor))
+
+            PjktFileExporter exporter = new PjktFileExporter(boothDescriptor.currentCommunity);
+            packagePath = exporter.CreateBoothfile(boothDescriptor.gameObject);
+
+            if (string.IsNullOrEmpty(packagePath))
             {
-                if (boothDescriptor.currentCommunity == "Debug")
-                {
-                    PjktSdkWindow.Notify("Created Debug Booth. No upload necessary.");
-                    AssetDatabase.Refresh();
-                    return;
-                }
-                
-                success = await UploadBoothToServer(boothDescriptor);
+                //failed to create zip for some reason
+                PjktSdkWindow.Notify($"Failed to create booth package. Ask for help on the discord.", BoothErrorType.Error);
+                return;
             }
             
+            if (boothDescriptor.currentCommunity == "Debug")
+            {
+                PjktSdkWindow.Notify("Created Debug Booth. No upload necessary.");
+                //AssetDatabase.Refresh();
+                return;
+            }
+                
+            success = await UploadBoothToServer(boothDescriptor);
             
             CleanupOperations(boothDescriptor);
             if (success) PjktSdkWindow.Notify("Booth Uploaded Successfully. See you at the event!");
         }
 
+        /*
         private static async Task<bool> CreateBoothPackage(BoothDescriptor boothDescriptor)
         {
             prefabPath = defaultPrefabPath + boothDescriptor.currentCommunity +"/" + boothDescriptor.boothName + ".prefab";
@@ -93,7 +110,7 @@ namespace PJKT.SDK2
                 PjktSdkWindow.Notify($"Booth upload failed, Couldn't create booth package\n {e}", BoothErrorType.Error);
                 return false;
             }
-        }
+        }*/
 
         private static async Task<bool> UploadBoothToServer(BoothDescriptor booth)
         {
@@ -108,7 +125,7 @@ namespace PJKT.SDK2
                 MultipartFormDataContent form = new MultipartFormDataContent();
                 form.Add(new StringContent(PjktEventManager.SelectedProjekt.id.ToString()), "project_id");
                 form.Add(new StringContent(Authentication.ActiveUser.GetCommunityId(booth.currentCommunity).ToString()), "community_id");
-                form.Add(new ByteArrayContent(fileBytes), "booth", booth.boothName + ".unitypackage");
+                form.Add(new ByteArrayContent(fileBytes), "booth", booth.boothName + ".zip");
                 form.Add(new ByteArrayContent(previewImageBytes), "preview", booth.boothName + " preview.png");
                 
                 
@@ -118,7 +135,7 @@ namespace PJKT.SDK2
                 HttpClientHandler handler = new HttpClientHandler() { CookieContainer = container };
                 HttpClient client = new HttpClient(handler);
 
-                byte[] bytes = await form.ReadAsByteArrayAsync();
+                //byte[] bytes = await form.ReadAsByteArrayAsync();
                 HttpResponseMessage response = await client.PostAsync(PJKTNet.defaultHost + "/booth/submit", form);
                 
                 if (response.StatusCode == HttpStatusCode.OK) return true;
@@ -135,18 +152,26 @@ namespace PJKT.SDK2
             }
         }
 
+        
         private static void CleanupOperations(BoothDescriptor booth)
         {
+            //really just need to remove the zip now
+            if (string.IsNullOrEmpty(packagePath)) return;
+            if (!File.Exists(packagePath)) return;
+            if (Path.GetExtension(packagePath) != ".zip") return; //dont delete random stuff pls
+            File.Delete(packagePath);
+
+            /*
             //Delete the temporary files and folders
             if (File.Exists(prefabPath)) File.Delete(prefabPath);
             string prefabFolderPath = prefabPath.TrimEnd('.','p','r','e','f','a','b');
             if (File.Exists(prefabFolderPath + ".meta")) File.Delete(prefabFolderPath + ".meta");
             
             if (File.Exists(packagePath)) File.Delete(packagePath);
-            string packagefolderPath = packagePath.TrimEnd('.','u','n','i','t','y','p', 'a', 'c', 'k', 'a', 'g', 'e');
-            if (File.Exists(packagefolderPath + ".meta")) File.Delete(packagefolderPath + ".meta");
+            //string packagefolderPath = packagePath.TrimEnd('.','u','n','i','t','y','p', 'a', 'c', 'k', 'a', 'g', 'e');
+            //if (File.Exists(packagefolderPath + ".meta")) File.Delete(packagefolderPath + ".meta");
             
-            if (UnityEngine.Windows.Directory.Exists(Path.GetDirectoryName(prefabFolderPath))) UnityEngine.Windows.Directory.Delete(Path.GetDirectoryName(prefabFolderPath));
+            if (UnityEngine.Windows.Directory.Exists(Path.GetDirectoryName(prefabFolderPath))) Directory.Delete(Path.GetDirectoryName(prefabFolderPath));
             if (UnityEngine.Windows.Directory.Exists(Path.GetDirectoryName(prefabFolderPath)  + ".meta")) Directory.Delete(Path.GetDirectoryName(prefabFolderPath) + ".meta");
             if (UnityEngine.Windows.File.Exists(defaultPrefabPath + booth.currentCommunity +".meta")) File.Delete(defaultPrefabPath + booth.currentCommunity +".meta");
 
@@ -154,9 +179,9 @@ namespace PJKT.SDK2
             packagePath = "";
             
             //Refresh the asset database
-            AssetDatabase.Refresh();
+            AssetDatabase.Refresh();*/
         }
-        
+        /*
         public static void CreatePath(string file)
         {
             if (Directory.Exists(file)) return;
@@ -167,6 +192,6 @@ namespace PJKT.SDK2
         public static void CreateDirectory(string file)
         {
             CreatePath(Path.GetDirectoryName(file));
-        }
+        }*/
     }
 }
